Rewrote the section with MongoDB data set.
The analysis done last year was based on an erroneously extracted
data set. The timeseries used are a merge of multiple different
HW and Sw configurations.

After this I also intend to add a section with the analysis Ishita did.
But I wanted to checkpoint and commit now, as everything else is done.
This version could be used as a paper if need be.